+++ lib/karafka/web/pro/commanding/commands/quiet.rb
@@ -1,0 +1,31 @@
+# frozen_string_literal: true
+
+# This Karafka component is a Pro component under a commercial license.
+# This Karafka component is NOT licensed under LGPL.
+#
+# All of the commercial components are present in the lib/karafka/pro directory of this
+# repository and their usage requires commercial license agreement.
+#
+# Karafka has also commercial-friendly license, commercial support and commercial components.
+#
+# By sending a pull request to the pro components, you are agreeing to transfer the copyright of
+# your code to Maciej Mensfeld.
+
+module Karafka
+  module Web
+    module Pro
+      module Commanding
+        # Namespace for commands the process can react to
+        module Commands
+          # Sends a signal to quiet the consumer
+          class Quiet < Base
+            # Performs the command
+            def call
+              ::Process.kill('TSTP', ::Process.pid)
+            end
+          end
+        end
+      end
+    end
+  end
+end
