+++ spec/lib/karafka/web/pro/commanding/listener_spec.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+RSpec.describe_current do
+  pending
+end
